plugin: remove unused setup_context functionality
It was previously possible to delegate "setup-context" to plugins, which
in turn were responsible to inject tokens into the lexer stream
that would drive the plugin specific grammar to the appropriate branch.

As it turned out all plugins just used the default functionality provided
by the core, which was simply injecting the plugin-type (source, destination,
etc) as a token. So the setup-context functionality was
unused completely.

This patch removes that.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>